Why I chose maps & a movie over machine learning
Above is an explanation of why, after all that effort put into my machine learning model, I decided to switch to a multimedia and interactive map approach.  I could have made a model with 90% accuracy and told you that it was legit.  In fact, I did make a model with 100% accuracy.  Hell, I could give you a model with any percent you want.  82% sounds good.  But then I'd be lying to you, and even though you'd most likely never realize it, I try to save up my yearly quota of lies for Christmas time around the kids.  And when my wife asks me if I want to watch a make up tutorial on YouTube.  So I gotta save those up.  It's only July after all.    
But the movie and the map are cool!
https://ripetimes.press/cracking-california-earthquake-project